% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Calculate a background of a data set}
\usage{
background(y, width = 20)
}
\arguments{
\item{y}{array, the data to be approximated}

\item{width}{float the half width of the window}
}
\value{
array of new points
}
\description{
Calculate a background of a data set
}
\details{
Calculate a background of a data set assuming this background
is a lower envelop of the curve represented in the 'y' array.
The process works by taking the minimum between the actual data
point and the average of two points at the distance of +/- j,
thus employing a window of 2j +1.
This procedure is repeated for j = 1 ... width.
}
\references{
Based on the work of:
Miroslav Morhác and Vladislav Matousek,
Peak clipping algorithms for background estimation in
spectroscopic data
Applied Spectroscopy 62(1): 91 - 106 (2008)
}
