% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{convolve.fft}
\alias{convolve.fft}
\title{convolution using FFT with nonzero padding}
\usage{
convolve.fft(x, y, padding = "", is.filter = TRUE)
}
\arguments{
\item{x, y}{float arrays to be convolved}

\item{padding}{int|text the padding would be len(x)+len(y) by default
if padding is 'full', then go up to the next
power of 2 in length}

\item{is.filter}{bool if TRUE, we do non-zero padding for filter application
y is then the filter kernel}
}
\value{
the convolved array with length of x
}
\description{
convolution using FFT with nonzero padding
}
\details{
When convolution filter is used for smoothing data, data and the
kernel are zero padded traditionally. However, this causes a cut-off
at the edges, which is often an undesired effect.

This version pads the first function with its first and last value
elminating cut off to 0 if these are non zero.
This way a smoothing will work better, but such a treatment should not
be done for other convolution operations, like autocorrelation.
Thus, the switch is.filter activates this feature.

One more word on padding: the signal is padded in a symmetric manner,
the kernel only to its end. Then the resulted range is cut out of the
returned convolved data.
Only the real part of the complex result is returned
(we assume x and y are real)
}
