% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{diff.gauss}
\alias{diff.gauss}
\title{differentiate data using a Gaussian kernel}
\usage{
\method{diff}{gauss}(y, sigma = 10, R = NULL)
}
\arguments{
\item{y}{the data vector}

\item{sigma}{width of the Gaussian (equivalent to
the standard deviation of a normal distribution)}

\item{R}{half width of the window used in the convolution
Window size is: 2R+1}
}
\value{
the differentiated vector with the same length as y
}
\description{
differentiate data using a Gaussian kernel
}
\details{
It is an envelope function calling the smooth.Gauss with deriv=1
}
