% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{find.peaks}
\alias{find.peaks}
\title{Find peaks in a dataset}
\usage{
find.peaks(
  a,
  sigma = 3,
  R = NULL,
  peak.threshold = 0.05,
  index.only = FALSE,
  peak.window = 2,
  both = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{a}{two column array, [x,y]}

\item{sigma}{width for the differential kernel}

\item{R}{window radius for the differential kernel (optional)}

\item{peak.threshold}{float a relative minimum height of the peaks
(compared to the maximum of y)}

\item{index.only}{boolean, if TRUE return the index of the peaks}

\item{peak.window}{int, how many points are taken before and after
the peak to refine its position as a sum(x*y)/sum(y)
and find the maximum of y within this range.}

\item{both}{boolean, if set return a list with pos and index}

\item{verbose, }{boolean, if set plot the derivative and selections}
}
\value{
an array of peak posiions in X or a list with index, position
         and peak height arrays (y.max). If indx.only is set, return
         an array of index positions
}
\description{
Find peaks in a dataset
}
\details{
find fast and dirty peak candidates, and their positions.
The idea is to take the first derivative using a Gaussian differential
kernel, and then find the zero transitions from positive to negative
slopes.
We check if peaks are closer than peak.window, and take the maximum
within this range (cutting secondary peaks off).
The function can return the index of a peak, its weighted position and
the hight of the local maximum.
}
