% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{Gauss.Kernel}
\alias{Gauss.Kernel}
\title{generate a 1D Gaussian kernel
using the equation 1/sqrt(2*pi)/sigma*exp(-(x)**2/(2*sigma**2))
Alternatively, use the derivative function that is -x/sigma**2 times
the Gaussian above.}
\usage{
Gauss.Kernel(sigma, R = NULL, deriv = 0)
}
\arguments{
\item{sigma}{float, standard deviation}

\item{R}{integer, number of points; if NULL, then use 3*sigma}

\item{deriv}{integer 0 or 1, the order of derivative to be used}
}
\value{
array containing the Gaussian
}
\description{
generate a 1D Gaussian kernel
using the equation 1/sqrt(2*pi)/sigma*exp(-(x)**2/(2*sigma**2))
Alternatively, use the derivative function that is -x/sigma**2 times
the Gaussian above.
}
