% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{integ.trap}
\alias{integ.trap}
\title{calculate a simple trapezoid formula numerical integral}
\usage{
integ.trap(x, y)
}
\arguments{
\item{x, y}{arrays of the same length}
}
\value{
the integral
}
\description{
calculate a simple trapezoid formula numerical integral
}
