% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{lin.background}
\alias{lin.background}
\title{background correction to remove linear backgrounds
Data sets often have a linear drift. To estimate this drift, the
lower part of the y data are taken into account and fitted using lm()
This background can be then subtracted.}
\usage{
lin.background(x, y = NULL, q = 0.5, verbose = FALSE)
}
\arguments{
\item{x, }{y array of data}

\item{y}{array or NULL the second column if x is single column}

\item{q}{float, a quantile unter which the values are used for fitting
by default it is 0.5 -> median}

\item{verbose}{Boolean, produce a plot showing the background and the fit}
}
\value{
returns a list with the new values, x,y and the fit object ft
}
\description{
background correction to remove linear backgrounds
Data sets often have a linear drift. To estimate this drift, the
lower part of the y data are taken into account and fitted using lm()
This background can be then subtracted.
}
