% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalProcessing.R
\name{read.all.files}
\alias{read.all.files}
\title{apply a reader function to a list of filenames found by dir()}
\usage{
read.all.files(
  folder = "./",
  pattern = ".*\\\\.csv$",
  reader = read.table,
  prefix = "",
  remove = "",
  remove.extension = TRUE,
  remove.dir = TRUE,
  recursive = FALSE,
  ...
)
}
\arguments{
\item{folder}{text, the folder to search in}

\item{pattern}{text, the file name pattern to load}

\item{reader}{function, returns the object read}

\item{prefix}{text, prepend this to the file names}

\item{remove}{text, pattern to be removed from the names}

\item{remove.extension}{Boolean, remove the last dot and its tail}

\item{remove.dir}{Boolean strip path from filenames}

\item{recursive}{Boolean, run into subfolders or not}

\item{...}{all further parameters are passed to the reader function}
}
\value{
nothing
}
\description{
apply a reader function to a list of filenames found by dir()
}
\details{
find a list of files based on folder and pattern via dir(),
apply reader(path, ...) to read them
convert their names to variable names using prefix and remove patterns
add them to the global environment
in all file names convert - or _ characters to .-s, and turn multiple '.'
to single ones.
}
