% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{smooth.Gauss}
\alias{smooth.Gauss}
\title{A function to smooth the data array y using a Gaussian kernel
It uses an offsetted padding and the convolve.fft function
to calculate the result.
The mean is subtracted from y for removing an offset.
If y is only smoothed, this mean is readded at the end. For the
derivative it is left out.}
\usage{
smooth.Gauss(y, sigma = 10, R = NULL, deriv = 0)
}
\arguments{
\item{y}{the data}

\item{sigma}{width of the Gaussian, standard deviation}

\item{R}{the half window size for the kernel; if NULL,
use automatic, 3*sigma see Gauss.Kernel}

\item{deriv}{derivative, possible values 0 or 1}
}
\value{
array of smoothed data
}
\description{
If R is not more than 3*sigma, the kernel is renormalized to avoid
offsetting the data.
}
