% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{background.gauss}
\alias{background.gauss}
\title{Calculate a background based on smoothing the data}
\usage{
background.gauss(y, sigma = 10, N = 10)
}
\arguments{
\item{y}{the original curve}

\item{sigma}{the standard deviation / width of the Gaussian
in points / index values, not X}

\item{N}{integer the number of runs}
}
\value{
an array of the background data
}
\description{
Calculate a background based on smoothing the data
}
\details{
Run the filter with sigma standard deviation N times,
and in every keep the lower value between the filtered
curve and the one before this filter step.
The resuls is the smoothed curve.
}
