% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{find a lower envelope for the function using a window of
2*width + 1, thus for every points +/- width
Based on the work of:
Miroslav Morhác and Vladislav Matousek,
Peak clipping algorithms for background estimation in
spectroscopic data
Applied Spectroscopy 62(1): 91 - 106 (2008)}
\usage{
background(y, width = 20)
}
\arguments{
\item{y}{array, the data to be approximated}

\item{width}{float the half width of the window}
}
\value{
array of new points
}
\description{
find a lower envelope for the function using a window of
2*width + 1, thus for every points +/- width
Based on the work of:
Miroslav Morhác and Vladislav Matousek,
Peak clipping algorithms for background estimation in
spectroscopic data
Applied Spectroscopy 62(1): 91 - 106 (2008)
}
