% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss-filters.R
\name{find.peaks}
\alias{find.peaks}
\title{find fast and dirty peak candidates, and their positions
it checks if peaks are closer than peak.window and cuts off
the second position in that case}
\usage{
find.peaks(
  a,
  sigma = 3,
  R = NULL,
  peak.threshold = 0.05,
  index.only = FALSE,
  peak.window = 2,
  both = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{a}{two column array, [x,y]}

\item{sigma}{width for the differential kernel}

\item{R}{window radius for the differential kernel (optional)}

\item{peak.threshold}{float a relative minimum height of the peaks
(compared to the maximum of y)}

\item{index.only}{boolean, if TRUE return the index of the peaks}

\item{peak.window}{int, how many points are taken before and after
the peak to refine its position as a sum(x*y)/sum(y)
and find the maximum of y within this range.}

\item{both}{boolean, if set return a list with pos and index}

\item{verbose, }{boolean, if set plot the derivative and selections}
}
\value{
an array of peak posiions in X or a list with index, position
         and peak height arrays (y.max). If indx.only is set, return
         an array of index positions
}
\description{
find fast and dirty peak candidates, and their positions
it checks if peaks are closer than peak.window and cuts off
the second position in that case
}
